move OffsetTrackingOp out of coregistration
