+++ s1tbx-op-insar-ui/src/main/java/org/esa/s1tbx/insar/gpf/ui/OffsetTrackingOpUI.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, see http://www.gnu.org/licenses/
  */
-package org.esa.s1tbx.insar.gpf.ui.coregistration;
+package org.esa.s1tbx.insar.gpf.ui;
 
 import org.esa.snap.core.datamodel.MetadataElement;
 import org.esa.snap.core.dataop.resamp.ResamplingFactory;
