+++ s1tbx-op-insar/src/main/resources/META-INF/services/org.esa.snap.core.gpf.OperatorSpi
@@ -2,7 +2,7 @@ org.esa.s1tbx.insar.gpf.coregistration.CreateStackOp$Spi
 org.esa.s1tbx.insar.gpf.coregistration.CrossCorrelationOp$Spi
 org.esa.s1tbx.insar.gpf.coregistration.WarpOp$Spi
 org.esa.s1tbx.insar.gpf.coregistration.DEMAssistedCoregistrationOp$Spi
-org.esa.s1tbx.insar.gpf.coregistration.OffsetTrackingOp$Spi
+org.esa.s1tbx.insar.gpf.OffsetTrackingOp$Spi
 org.esa.s1tbx.insar.gpf.GoldsteinFilterOp$Spi
 org.esa.s1tbx.insar.gpf.PhaseToElevationOp$Spi
 org.esa.s1tbx.insar.gpf.StackAveragingOp$Spi
